refactor: 이전에 참여한 소모임, 스크랩한 소모임 마이페이지로 연결
